+++ built/jtree.node.d.ts
@@ -4,5 +4,6 @@ declare class jtreeNode extends jtree {
     static executeFile: (programPath: string, grammarPath: string) => Promise<any>;
     static makeProgram: (programPath: string, grammarPath: string) => AbstractRuntimeProgram;
     static getProgramConstructor: (grammarPath: string) => import("./grammar/AbstractRuntimeProgramConstructorInterface").default;
+    static combineFiles: (globPatterns: string[]) => import("./base/TreeNode").default;
 }
 export default jtreeNode;